requirements.txt
flask
pdfplumber
pytesseract
pillow
